requirements.txt
aiohttp==3.7.4.post0
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
certifi==2021.10.8
cffi==1.14.3
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.12
entrypoints==0.3
gitdb==4.0.7
GitPython==3.1.18
idna==2.10
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.1
Jinja2==2.11.2
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kiwisolver==1.3.2
lizard==1.17.7
MarkupSafe==1.1.1
matplotlib==3.4.3
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==5.1.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.2
notebook==6.1.4
numpy==1.19.2
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
pip==21.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycparser==2.20
PyDriller==2.0
PyGithub==1.55
Pygments==2.7.2
PyJWT==2.1.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
requests==2.25.1
scikit-learn==1.0.1
scipy==1.6.2
Send2Trash==1.5.0
setuptools==52.0.0.post20210125
sip==4.19.13
six==1.15.0
sklearn==0.0
smmap==4.0.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==3.0.0
tornado==6.0.4
traitlets==5.0.5
typing-extensions==3.10.0.2
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.36.2
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.0
